back-end/pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["RanaReebaal <rebaalhussain@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pytest = "^8.3.3"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
